+++ app/Actions/Database/DeleteDatabaseUser.php
@@ -9,8 +9,12 @@ class DeleteDatabaseUser
 {
     public function delete(Server $server, DatabaseUser $databaseUser): void
     {
+        $service = $server->database();
+        if (! $service instanceof \App\Models\Service) {
+            throw new \Exception('Database service not found');
+        }
         /** @var \App\SSH\Services\Database\Database $handler */
-        $handler = $server->database()->handler();
+        $handler = $service->handler();
         $handler->deleteUser($databaseUser->username, $databaseUser->host);
         $databaseUser->delete();
     }
