+++ app/Actions/Queue/EditQueue.php
@@ -29,8 +29,12 @@ class EditQueue
         $queue->save();
 
         dispatch(function () use ($queue): void {
+            $service = $queue->server->processManager();
+            if (! $service) {
+                throw new \Exception('Process manager service not found');
+            }
             /** @var ProcessManager $processManager */
-            $processManager = $queue->server->processManager()->handler();
+            $processManager = $service->handler();
             $processManager->delete($queue->id, $queue->site_id);
 
             $processManager->create(
